Defer the destruction of gameobjects
Defering the destruction of game objects until after the frame means
that bugs wehre an object is destroyed and you access it in the same
frame happen less.

A developer will still need to clean up after themselves for the next
frame but it's not as confusing as objects being destroyed right away
and losing the ability to access them. This way we can have one guard
at the start of the frame and know we are safe for the duration of that
frame to access the data of the object.