Add a clickbox for ui sprites
This can be used to get certain info about the mouse and interactions
with the UISprite renderer such as clicks, pointer entering, exiting
or if it's inside. This can be used for the basis of a button and
works with different shaped UI renderers with different rotations,
positions or scales.